Improve discover (#296)
* Compartmentalize discover pt 1

* Compartmentalize discover pt 3

* Improve discover hit buttons

* Extract filters bar component

* Extract filters component to its own file

* Fix search snippets and small design issues
