+++ site/gatsby-site/src/components/WebArchiveLink.js
@@ -10,7 +10,7 @@ async function getSnapshotURL(url, date) {
   return response.archived_snapshots.closest.url.replace('http:', 'https:');
 }
 
-export default function WebArchiveLink({ url, date, children }) {
+export default function WebArchiveLink({ url, date, children, className }) {
   const onClick = async (e) => {
     e.preventDefault();
     const win = window.open('', '_blank');
@@ -25,7 +25,7 @@ export default function WebArchiveLink({ url, date, children }) {
   };
 
   return (
-    <a onClick={onClick} href={url} target="_blank" rel="noopener noreferrer">
+    <a className={className} onClick={onClick} href={url} target="_blank" rel="noopener noreferrer">
       {children}
     </a>
   );
